+++ src/main/java/co/nvqa/common/core/client/ReservationClient.java
@@ -92,15 +92,15 @@ public class ReservationClient extends SimpleApiClient {
   }
 
   public ReservationResponse updatePriorityLevelOfReservation(long pickupAddressId,
-      long globalShipperId, long legacyShipperId,
+      long globalShipperId,
       long priorityLevel, long reservationId) {
     String url = "reservation/2.0/reservations/{reservation_id}";
 
     RequestSpecification spec = createAuthenticatedRequest()
         .pathParam("reservation_id", reservationId)
         .body(
-            f("{\"pickup_address_id\":%d,\"global_shipper_id\":%d,\"legacy_shipper_id\":%d,\"reservation_type_value\":0,\"priority_level\":%d}",
-                pickupAddressId, globalShipperId, legacyShipperId, priorityLevel));
+            f("{\"pickup_address_id\":%d,\"global_shipper_id\":%d,\"reservation_type_value\":0,\"priority_level\":%d}",
+                pickupAddressId, globalShipperId, priorityLevel));
 
     Response r = doPost("RESERVATION - UPDATE PRIORITY LEVEL", spec, url);
     r.then().contentType(ContentType.JSON);